Moved cfor syntax and tests to the kernel package.
