+++ kernel-macros/src/main/scala_2.10/spire/macros/compat.scala
