+++ kernel-macros/src/main/scala_2.11/spire/macros/compat.scala
