+++ kernel-macros/src/main/scala_2.12/spire/macros/compat.scala
