+++ kernel-macros/src/main/scala_2.13/spire/macros/compat.scala
