+++ kernel/shared/src/main/scala/spire/syntax/cfor.scala
@@ -1,0 +1,3 @@
+package spire.syntax
+
+object cfor extends CforSyntax
